+++ frontend/src/app/components/containers/admin/user-management/select-user.tsx
@@ -1,34 +1,17 @@
-import { Dispatch, FC, Fragment, SetStateAction, useEffect, useState } from "react";
+import { Dispatch, FC, SetStateAction } from "react";
 import { Button } from "react-bootstrap";
-import { useAppDispatch, useAppSelector } from "@hooks/hooks";
-import { assignOfficerToOffice, selectOfficers, selectOfficersDropdown } from "@store/reducers/officer";
+import { useAppSelector } from "@hooks/hooks";
+import { selectOfficers } from "@store/reducers/officer";
 import { CompSelect } from "@components/common/comp-select";
 import Option from "@apptypes/app/option";
-import { fetchOfficeAssignments, selectOfficesForAssignmentDropdown, selectOffices } from "@store/reducers/office";
-import { ToastContainer } from "react-toastify";
-import { ToggleError, ToggleSuccess } from "@common/toast";
-import { clearNotification, selectNotification } from "@store/reducers/app";
-import { selectAgencyDropdown, selectTeamDropdown } from "@store/reducers/code-table";
-import { ROLE_OPTIONS } from "@constants/ceeb-roles";
-import { generateApiParameters, get, patch } from "@common/api";
-import config from "@/config";
 import { Officer } from "@apptypes/person/person";
-import { UUID } from "crypto";
-import { ValidationMultiSelect } from "@common/validation-multiselect";
 import "@assets/sass/user-management.scss";
-import { AsyncTypeahead, Typeahead } from "react-bootstrap-typeahead";
-import Select, { OptionProps, StylesConfig, components } from "react-select";
 
 interface SelectUserProps {
   officer: any;
   officerError: string;
-  userIdirs: any;
-  officerGuid: any;
   setOfficer: Dispatch<SetStateAction<Option | undefined>>;
   setOfficerError: Dispatch<SetStateAction<string>>;
-  getUserIdir: (person_guid: string, lastName: string, firstName: string) => Promise<void>;
-  setSelectedUserIdir: Dispatch<SetStateAction<string>>;
-  updateUserIdirByOfficerId: (userIdir: string, officerGuid: string) => Promise<void>;
   handleEdit: () => void;
   handleAddNewUser: () => void;
 }
@@ -39,10 +22,6 @@ export const SelectUser: FC<SelectUserProps> = ({
   handleAddNewUser,
   officer,
   officerError,
-  userIdirs,
-  setSelectedUserIdir,
-  updateUserIdirByOfficerId,
-  officerGuid,
   handleEdit,
 }) => {
   const officers = useAppSelector(selectOfficers);
